fix gcc 4.1 retardation

+++ trunk/test/speedtest_msg.c
@@ -59,10 +59,10 @@ size_t test_msgpk_encode(char* dest)
 {
   modp_msgpk_ctx ctx;
   modp_msgpk_init(&ctx, dest);
-  modp_msgpk_map_open(&ctx, 4);
+  modp_msgpk_map_open(&ctx, (size_t)4);
 
   modp_msgpk_add_cstring(&ctx, "start_ms");
-  modp_msgpk_add_uint32(&ctx, 123456789);
+  modp_msgpk_add_uint32(&ctx, (uint32_t) 123456789);
 
   modp_msgpk_add_cstring(&ctx, "remote_ip");
   modp_msgpk_add_cstring(&ctx, "123.123.123.13");
@@ -71,24 +71,24 @@ size_t test_msgpk_encode(char* dest)
   modp_msgpk_add_cstring(&ctx, "GET /foobar HTTP/1.1");
 
   modp_msgpk_add_cstring(&ctx, "headers_in");
-  modp_msgpk_ary_open(&ctx, 3);
+  modp_msgpk_ary_open(&ctx, (size_t)3);
 
-  modp_msgpk_ary_open(&ctx, 2);
+  modp_msgpk_ary_open(&ctx, (size_t)2);
   modp_msgpk_add_cstring(&ctx, "Accept");
   modp_msgpk_add_cstring(&ctx, "*/*");
   modp_msgpk_ary_close(&ctx);
 
-  modp_msgpk_ary_open(&ctx, 2);
+  modp_msgpk_ary_open(&ctx, (size_t)2);
   modp_msgpk_add_cstring(&ctx, "Content-type");
   modp_msgpk_add_cstring(&ctx, "text/plain");
   modp_msgpk_ary_close(&ctx);
 
-  modp_msgpk_ary_open(&ctx, 2);
+  modp_msgpk_ary_open(&ctx, (size_t)2);
   modp_msgpk_add_cstring(&ctx, "Connection");
   modp_msgpk_add_cstring(&ctx, "close");
   modp_msgpk_ary_close(&ctx);
 
-  modp_msgpk_ary_open(&ctx, 2);
+  modp_msgpk_ary_open(&ctx, (size_t)2);
   modp_msgpk_add_cstring(&ctx, "User-agent");
   modp_msgpk_add_cstring(&ctx, "Mozilla/5.0 (iPad; U; CPU OS 3_2_1 like Mac OS X; en-us) AppleWebKit/531.21.10 (KHTML, like Gecko) Mobile/7B405");
   modp_msgpk_ary_close(&ctx);
